Aggiornamento Refactoring dbandbusiness gruppo 6(Francesco Citarella, Andrea Siboni, Cesare Colella, Riccardo Brunelli)
