+++ DataLayer/StudentData.cs
@@ -1,0 +1,1455 @@
+﻿using System;
+using System.Data.Common;
+using System.Data;
+using System.Data.SQLite;
+using System.IO;
+using System.Collections.Generic;
+using gamon;
+using System.Windows.Forms;
+using SchoolGrades.DbClasses;
+
+namespace SchoolGrades.DataLayer
+{
+    //Cesare Colella, Francesco Citarella, Andrea Siboni, Riccardo Brunelli 4L
+
+    class StudentData
+    {
+        DataLayer dl = new DataLayer();
+        internal DataTable GetGradesOfStudent(Student Student, string SchoolYear,
+            string IdGradeType, string IdSchoolSubject,
+            DateTime DateFrom, DateTime DateTo)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT Grades.idGrade,datetime(Grades.timeStamp), Students.idStudent," +
+                "lastName,firstName," +
+                " Grades.value AS 'grade', Grades.weight," +
+                " Grades.idGradeParent" +
+                " FROM Grades" +
+                " JOIN Students" +
+                " ON Students.idStudent=Grades.idStudent" +
+                " JOIN Classes_Students" +
+                " ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Students.idStudent =" + Student.IdStudent +
+                " AND Grades.idSchoolYear='" + SchoolYear + "'" +
+                " AND Grades.idGradeType = '" + IdGradeType + "'" +
+                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
+                " AND Grades.Value > 0" +
+                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
+                " ORDER BY lastName, firstName, Students.idStudent, Grades.timestamp Desc;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        internal object GetGradesWeightsOfStudentOnOpenGrades(Student currentStudent, string stringKey1, string stringKey2, DateTime value1, DateTime value2)
+        {
+            throw new NotImplementedException();
+
+        }
+
+        internal int? UpdateAnnotationGroup(StudentAnnotation currentAnnotation, Student currentStudent)
+        {
+            throw new NotImplementedException();
+        }
+
+
+
+
+        internal DataTable GetGradesWeightsOfClassOnOpenGrades(Class Class,
+            string IdGradeType, string IdSchoolSubject, DateTime DateFrom, DateTime DateTo)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                // find the macro grade type of the micro grade
+                // TODO take it from a Grade passed as parameter 
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT idGradeTypeParent " +
+                    "FROM GradeTypes " +
+                    "WHERE idGradeType='" + IdGradeType + "'; ";
+                string idGradeTypeParent = (string)cmd.ExecuteScalar();
+
+                string query = "SELECT Grades.idGrade,Students.idStudent,lastName,firstName" +
+                ",SUM(Grades.weight)/100 AS 'GradesFraction', 1 - SUM(Grades.weight)/100 AS LeftToCloseAssesments" +
+                ",COUNT() AS 'GradesCount'" +
+                " FROM Grades, Grades AS Parents " +
+                " JOIN Classes_Students ON Students.idStudent=Grades.idStudent" +
+                " JOIN Students ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass =" + Class.IdClass +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                " AND (Grades.idGradeType='" + IdGradeType + "'" +
+                " OR Grades.idGradeType IS NULL)" +
+                " AND Grades.idSchoolSubject='" + IdSchoolSubject + "'" +
+                " AND Grades.value IS NOT NULL AND Grades.value <> 0" +
+                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
+                " AND Parents.idGradeType = '" + idGradeTypeParent + "'" +
+                " AND Grades.idGradeParent = Parents.idGrade" +
+                " AND (Parents.Value is null or Parents.Value = 0)" +
+                " AND NOT Students.disabled" +
+                " GROUP BY Students.idStudent" +
+                " ORDER BY GradesFraction ASC, lastName, firstName, Students.idStudent;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        private int NextKey(string Table, string Id)
+        {
+            int nextId;
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT MAX(" + Id + ") FROM " + Table + ";";
+                var firstColumn = cmd.ExecuteScalar();
+                if (firstColumn != DBNull.Value)
+                {
+                    nextId = int.Parse(firstColumn.ToString()) + 1;
+                }
+                else
+                {
+                    nextId = 1;
+                }
+                cmd.Dispose();
+            }
+            return nextId;
+        }
+
+
+        internal void DeleteValueOfGrade(int IdGrade)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+
+                cmd.CommandText = "UPDATE Grades" +
+                           " Set" +
+                           " value=null" +
+                           " WHERE IdGrade=" + IdGrade +
+                           ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+
+        internal void ToggleDisableOneStudent(int? idStudent)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+
+                cmd.CommandText = "UPDATE Students" +
+                           " Set" +
+                           " disabled = ~disabled" +
+                           " WHERE IdStudent =" + idStudent +
+                           ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        internal void EraseStudentsPhoto(int? IdStudent, string SchoolYear)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM StudentsPhotos_Students" +
+                    " WHERE idStudent=" + IdStudent +
+                    " AND idSchoolYear='" + SchoolYear + "'" +
+                    ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        internal DataTable GetUnfixedGrades(Student Student, string IdSchoolSubject,
+            double Threshold)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT Grades.IdGrade,Grades.idStudent,Grades.value,Grades.timestamp,Grades.isFixed," +
+                    "Grades.idGradeType,Grades.idQuestion,Questions.text,Questions.*,Grades.*" +
+                    " FROM Grades" +
+                    " JOIN Questions ON Grades.idQuestion=Questions.idQuestion" +
+                    " WHERE idStudent=" + Student.IdStudent +
+                    " AND Grades.value<" + Threshold.ToString() +
+                    " AND (Grades.isFixed=0 OR Grades.isFixed is NULL)";
+                if (IdSchoolSubject != "")
+                    query += " AND Grades.idSchoolSubject='" + IdSchoolSubject + "'";
+                query += ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetUnfixedGradesInTheYear");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dSet.Dispose();
+                dAdapt.Dispose();
+            }
+            return t;
+        }
+
+        internal string GetFilePhoto(int? IdStudent, string SchoolYear)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                string query = "SELECT StudentsPhotos.photoPath" +
+                    " FROM StudentsPhotos_Students, StudentsPhotos" +
+                    " WHERE StudentsPhotos_Students.idStudentsPhoto = StudentsPhotos.idStudentsPhoto";
+                if (SchoolYear != null && SchoolYear != "")
+                    query += " AND StudentsPhotos_Students.idSchoolYear='" + SchoolYear + "'";
+                query += " AND StudentsPhotos_Students.idStudent = " + IdStudent + "; ";
+                string NamePath = null;
+                try
+                {
+                    cmd.CommandText = query;
+                    NamePath = (string)cmd.ExecuteScalar();
+                }
+                catch (Exception ex)
+                {
+
+                }
+                cmd.Dispose();
+                return NamePath;
+            }
+        }
+
+        internal void AddLinkToOldPhoto(int? IdStudent, string IdPreviousSchoolYear, string IdNextSchoolYear)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                // get the code of the previous photo
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT idStudentsPhoto" +
+                    " FROM StudentsPhotos_Students" +
+                    " WHERE idSchoolYear='" + IdPreviousSchoolYear + "'" +
+                    " AND StudentsPhotos_Students.idStudent = " + IdStudent + "; ";
+                int? idStudentsPhoto = (int?)cmd.ExecuteScalar();
+                if (idStudentsPhoto != null)
+                {
+                    // add link to old photo
+                    cmd = conn.CreateCommand();
+                    cmd.CommandText = "INSERT INTO StudentsPhotos_Students " +
+                    "(idStudent, idStudentsPhoto, idSchoolYear) " +
+                    "Values (" +
+                    "" + IdStudent + "" +
+                    "," + idStudentsPhoto + "" +
+                    ",'" + IdNextSchoolYear + "'" +
+                    ");";
+                    cmd.ExecuteNonQuery();
+                }
+                cmd.Dispose();
+            }
+        }
+
+        internal void SaveStudentsOfList(List<Student> studentsList, DbConnection conn)
+        {
+            foreach (Student s in studentsList)
+            {
+                SaveStudent(s, conn);
+            }
+        }
+
+        internal int? SaveStudent(Student Student, DbConnection conn)
+        {
+            bool leaveConnectionOpen = true;
+            if (conn == null)
+            {
+                conn = dl.Connect();
+                leaveConnectionOpen = false;
+            }
+            DbCommand cmd = conn.CreateCommand();
+            cmd.CommandText = "UPDATE Students " +
+                "SET" +
+                " idStudent=" + Student.IdStudent +
+                ",lastName='" + SqlVal.SqlString(Student.LastName) + "'" +
+                ",firstName='" + SqlVal.SqlString(Student.FirstName) + "'" +
+                ",residence='" + SqlVal.SqlString(Student.Residence) + "'" +
+                ",birthDate=" + SqlVal.SqlDate(Student.BirthDate.ToString()) + "" +
+                ",email='" + SqlVal.SqlString(Student.Email) + "'" +
+                //",schoolyear='" + SqlVal.SqlString(Student.SchoolYear) + "'" +
+                ",drawable='" + SqlVal.SqlBool(Student.Eligible) + "'" +
+                ",origin='" + SqlVal.SqlString(Student.Origin) + "'" +
+                ",birthPlace='" + SqlVal.SqlString(Student.BirthPlace) + "'" +
+                ",drawable=" + SqlVal.SqlBool(Student.Eligible) + "" +
+                ",disabled=" + SqlVal.SqlBool(Student.Disabled) + "" +
+                ",VFCounter=" + Student.RevengeFactorCounter + "" +
+                " WHERE idStudent = " + Student.IdStudent +
+                ";";
+            cmd.ExecuteNonQuery();
+            cmd.Dispose();
+            if (!leaveConnectionOpen)
+            {
+                conn.Close();
+                conn.Dispose();
+            }
+            return Student.IdStudent;
+        }
+
+
+
+        internal DataTable GetStudentsSameName(string LastName, string FirstName)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT Students.IdStudent, Students.lastName, Students.firstName," +
+                    " Classes.abbreviation, Classes.idSchoolYear" +
+                    " FROM Students" +
+                    " LEFT JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent " +
+                    " LEFT JOIN Classes ON Classes.idClass = Classes_Students.idClass " +
+                    " WHERE Students.lastName LIKE '%" + LastName + "%'" +
+                    " AND Students.firstName LIKE '%" + FirstName + "%'" +
+                    ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetStudentsSameName");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dSet.Dispose();
+                dAdapt.Dispose();
+            }
+            return t;
+        }
+
+        internal DataTable FindStudentsLike(string LastName, string FirstName)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                DataAdapter dAdapt;
+                DataSet dSet = new DataSet();
+                string query = "SELECT Students.IdStudent, Students.lastName, Students.firstName," +
+                    " Classes.abbreviation, Classes.idSchoolYear" +
+                    " FROM Students" +
+                    " LEFT JOIN Classes_Students ON Students.idStudent = Classes_Students.idStudent " +
+                    " LEFT JOIN Classes ON Classes.idClass = Classes_Students.idClass ";
+                if (LastName != "" && LastName != null)
+                {
+                    query += "WHERE Students.lastName LIKE '%" + LastName + "%'";
+                    if (FirstName != "" && FirstName != null)
+                    {
+                        query += " AND Students.firstName LIKE '%" + FirstName + "%'";
+                    }
+                }
+                else
+                {
+                    if (FirstName != "" && FirstName != null)
+                    {
+                        query += " WHERE Students.firstName LIKE '%" + FirstName + "%'";
+                    }
+                }
+                query += ";";
+                dAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                dSet = new DataSet("GetStudentsSameName");
+                dAdapt.Fill(dSet);
+                t = dSet.Tables[0];
+
+                dSet.Dispose();
+                dAdapt.Dispose();
+            }
+            return t;
+        }
+
+        internal void PutStudentInClass(int? IdStudent, int? IdClass)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                // add student to the class
+                DbCommand cmd = conn.CreateCommand();
+                // check if already present
+                cmd.CommandText = "SELECT IdStudent FROM Classes_Students " +
+                    "WHERE idClass=" + IdClass + " AND idStudent=" + IdStudent + "" +
+                ";";
+                if (cmd.ExecuteScalar() == null)
+                {
+                    // add student to the class
+                    cmd.CommandText = "INSERT INTO Classes_Students " +
+                    "(idClass, idStudent) " +
+                    "Values ('" + IdClass + "'," + IdStudent + "" +
+                    ");";
+                }
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        internal int CopyAndLinkOnePhoto(Student Student, Class Class, string PathAndFileName)
+        {
+            if (!File.Exists(PathAndFileName))
+            {
+                throw new FileNotFoundException(@"[" + PathAndFileName + " not found.]");
+                //return 0;
+            }
+            if (File.Exists(PathAndFileName + "TEMP"))
+            {
+                File.Delete(PathAndFileName + "TEMP");
+            }
+            File.Copy(PathAndFileName, PathAndFileName + "TEMP");
+
+            string ext = Path.GetExtension(PathAndFileName);
+            string newFileName = Commons.PathImages + "\\" + Class.SchoolYear + Class.Abbreviation + "\\" +
+                Student.LastName + "_" + Student.FirstName + "_" + Class.Abbreviation + Class.SchoolYear + ext;
+            if (!Directory.Exists(Commons.PathImages + "\\" + Class.SchoolYear + Class.Abbreviation + "\\"))
+            {
+                Directory.CreateDirectory(Commons.PathImages + "\\" + Class.SchoolYear + Class.Abbreviation + "\\");
+            }
+            if (File.Exists(newFileName))
+            {
+                // !!!! TODO: resolve the problem of the lock that is still active here, 
+                // !!!! despite many attempts to free it !!!!
+                File.Delete(newFileName);
+            }
+            File.Move(PathAndFileName + "TEMP", newFileName);
+
+            // trova la chiave per la prossima foto 
+            int codiceFoto = NextKey("StudentsPhotos", "idStudentsPhoto");
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+
+                // aggiunge la foto alle foto (cartella relativa, cui verrà aggiunta la path delle foto)
+                cmd.CommandText = "INSERT INTO StudentsPhotos " +
+                "(idStudentsPhoto, photoPath)" +
+                "Values " +
+                "('" + codiceFoto + "','" + SqlVal.SqlString(Class.SchoolYear) +
+                SqlVal.SqlString(Class.Abbreviation) + "\\" + SqlVal.SqlString(Student.LastName) + "_" + SqlVal.SqlString(Student.FirstName) +
+                "_" + SqlVal.SqlString(Class.Abbreviation) + SqlVal.SqlString(Class.SchoolYear) + ".jpg" +
+                "');";
+                cmd.ExecuteNonQuery();
+
+                // cancella tutti gli eventuali riferimenti nella tabella link delle foto ad altre foto nell'anno 
+                cmd.CommandText = "DELETE FROM StudentsPhotos_Students " +
+                    "WHERE idStudent=" + Student.IdStudent +
+                    " AND idSchoolYear='" + Class.SchoolYear + "'" +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // aggiunge questa foto alla tabella link delle foto
+                cmd.CommandText = "INSERT INTO StudentsPhotos_Students " +
+                    "(idStudentsPhoto, idStudent, idSchoolYear) " +
+                    "Values (" + codiceFoto + "," + Student.IdStudent + ",'" + SqlVal.SqlString(Class.SchoolYear) +
+                    "');";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            return codiceFoto;
+        }
+
+        
+
+        internal int CreateClassAndStudents(string[,] StudentsData, string ClassAbbreviation,
+                    string ClassDescription, string SchoolYear, string OfficialSchoolAbbreviation,
+                    bool LinkPhoto)
+        {
+            // creation of a new class in the Classes table
+
+            // finds a key for the new class
+            int idClass = NextKey("Classes", "idClass");
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Classes " +
+                    "(idClass, Desc, idSchoolYear, idSchool, abbreviation) " +
+                    "Values (" + idClass + ",'" + SqlVal.SqlString(ClassDescription) + "','" +
+                    SqlVal.SqlString(SchoolYear) + "','" + SqlVal.SqlString(OfficialSchoolAbbreviation) + "','" +
+                    SqlVal.SqlString(ClassAbbreviation) + "'" +
+                    ");";
+                cmd.ExecuteNonQuery();
+
+                // find the key for next student
+                int idNextStudent = NextKey("Students", "idStudent");
+                // find the key for next picture 
+                int idNextPhoto = NextKey("StudentsPhotos", "idStudentsPhoto");
+                // add the student to the students' table 
+                // start from the second row of the file, first row is descriptions 
+                for (int riga = 1; riga < StudentsData.GetLength(0); riga++)
+                {
+                    int rigap1 = riga + 1;
+                    // create new student
+                    cmd.CommandText = "INSERT INTO Students " +
+                        "(idStudent, lastName, firstName, residence, origin, email, birthDate, birthPlace) " +
+                        "Values (" +
+                        "'" + idNextStudent + "','" +
+                        SqlVal.SqlString(StudentsData[riga, 1]) + "','" +
+                        SqlVal.SqlString(StudentsData[riga, 2]) + "','" +
+                        SqlVal.SqlString(StudentsData[riga, 3]) + "','" +
+                        SqlVal.SqlString(StudentsData[riga, 4]) + "','" +
+                        SqlVal.SqlString(StudentsData[riga, 5]) + "'," +
+                        SqlVal.SqlDate(StudentsData[riga, 6]) + ",'" +
+                        SqlVal.SqlString(StudentsData[riga, 7]) + "'" +
+                        ");";
+                    cmd.ExecuteNonQuery();
+
+                    // aggiunge lo studente alla classe
+                    cmd.CommandText = "INSERT INTO Classes_Students " +
+                        "(idClass, idStudent, registerNumber) " +
+                        "Values ('" + idClass + "','" + idNextStudent + "','" + rigap1.ToString() + "'" +
+                        ");";
+                    cmd.ExecuteNonQuery();
+
+                    if (LinkPhoto)
+                    {
+                        // aggiunge la foto alle foto
+                        cmd.CommandText = "INSERT INTO StudentsPhotos " +
+                            "(idStudentsPhoto, photoPath)" +
+                            "Values " +
+                            "('" + idNextPhoto + "','" + SqlVal.SqlString(SchoolYear) +
+                            SqlVal.SqlString(ClassAbbreviation) + "\\" + SqlVal.SqlString(StudentsData[riga, 1]) + "_" +
+                            SqlVal.SqlString(StudentsData[riga, 2]) + "_" + SqlVal.SqlString(ClassAbbreviation) +
+                            SqlVal.SqlString(SchoolYear) + ".jpg" + // TODO mettere l'estensione del file che c'è effettivamente
+                            "');"; // relative path. Home path will be added at visualization time 
+                        cmd.ExecuteNonQuery();
+
+                        // add the picture to the link table
+                        cmd.CommandText = "INSERT INTO StudentsPhotos_Students " +
+                            "(idStudentsPhoto, idStudent, idSchoolYear) " +
+                            "Values (" + idNextPhoto + "," + idNextStudent + ",'" + SqlVal.SqlString(SchoolYear) +
+                            "');";
+                        cmd.ExecuteNonQuery();
+                        idNextPhoto++;
+                    }
+                    idNextStudent++;
+                }
+                cmd.Dispose();
+            }
+            return idClass;
+        }
+
+ 
+
+        internal int CreateStudent(Student Student)
+        {
+            // trova una chiave da assegnare al nuovo studente
+            int idStudent = NextKey("Students", "idStudent");
+            Student.IdStudent = idStudent;
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Students " +
+                    "(idStudent,lastName,firstName,residence,origin," +
+                    "email,birthDate,birthPlace,disabled) " +
+                    "Values ('" + Student.IdStudent + "','" +
+                    SqlVal.SqlString(Student.LastName) + "','" +
+                    SqlVal.SqlString(Student.FirstName) + "','" +
+                    SqlVal.SqlString(Student.Residence) + "','" +
+                    SqlVal.SqlString(Student.Origin) + "','" +
+                    SqlVal.SqlString(Student.Email) + "'," +
+                    SqlVal.SqlDate(Student.BirthDate.ToString()) + ",'" +
+                    SqlVal.SqlString(Student.BirthPlace) + "'," +
+                    "false" +
+                    ");";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            return idStudent;
+        }
+
+        internal int CreateStudentFromStringMatrix(string[,] StudentData, int? StudentRow)
+        {
+            // trova una chiave da assegnare al nuovo studente
+            int codiceStudente = NextKey("Students", "idStudent");
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "INSERT INTO Students " +
+                    "(idStudent, lastName, firstName, residence, origin, email," +
+                    //" birthDate," + // save also this field after SqlVal.SqlDate( will include the '', to better treat null values
+                    " birthPlace) " +
+                    "Values (" + codiceStudente + "," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 1]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 2]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 3]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 4]) + "'," +
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 5]) + "'," +
+                    //"'" + SqlVal.SqlDate(StudentData[(int)StudentRow, 6]) + "'," + // save also this field after SqlVal.SqlDate( will include the ''
+                    "'" + SqlVal.SqlString(StudentData[(int)StudentRow, 7]) + "'" +
+                    ");";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            return codiceStudente;
+        }
+
+      
+
+        /// <summary>
+        /// 
+        /// </summary>
+        /// <param name="IdClass">Id of the class to be searched</param>
+        /// <param name="conn">Connection already open on a database different from standard. 
+        /// If not null this connection is left open</param>
+        /// <returns>List of the </returns>
+        internal List<Student> GetStudentsOfClass(int? IdClass, DbConnection conn)
+        {
+            List<Student> l = new List<Student>();
+            bool leaveConnectionOpen = true;
+
+            if (conn == null)
+            {
+                conn = dl.Connect();
+                leaveConnectionOpen = false;
+            }
+            DbDataReader dRead;
+            DbCommand cmd = conn.CreateCommand();
+            string query = "SELECT Students.*" +
+                " FROM Students" +
+                " JOIN Classes_Students ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass=" + IdClass +
+            ";";
+            cmd.CommandText = query;
+            dRead = cmd.ExecuteReader();
+
+            while (dRead.Read())
+            {
+                Student s = GetStudentFromRow(dRead);
+                l.Add(s);
+            }
+            if (!leaveConnectionOpen)
+            {
+                conn.Close();
+                conn.Dispose();
+            }
+            return l;
+        }
+
+        internal Class GetClassOfStudent(string IdSchool, string SchoolYearCode, Student Student)
+        {
+            Class c = new Class();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT Classes.*" +
+                   " FROM Classes" +
+                   " JOIN Classes_Students ON Classes.idClass = Classes_Students.idClass" +
+                   " JOIN Students ON Students.idStudent = Classes_Students.idStudent" +
+                   " WHERE" +
+                   " Classes.idSchool = '" + SqlVal.SqlString(IdSchool) + "'" +
+                   " AND Classes.idSchoolYear = '" + SqlVal.SqlString(SchoolYearCode) + "'" +
+                   " AND Students.IdStudent = " + Student.IdStudent +
+                   ";";
+                dRead = cmd.ExecuteReader();
+
+                while (dRead.Read())
+                {
+                    GetClassFromRow(c, dRead);
+                    break; // just the first! 
+                }
+            }
+            return c;
+        }
+
+        private void GetClassFromRow(Class Class, DbDataReader Row)
+        {
+            if (Class == null)
+                Class = new Class();
+            Class.IdClass = (int)Row["idClass"];
+            Class.Abbreviation = SafeDb.SafeString(Row["abbreviation"]);
+            Class.IdSchool = SafeDb.SafeString(Row["idSchool"]);
+            Class.PathRestrictedApplication = SafeDb.SafeString(Row["pathRestrictedApplication"]);
+            Class.SchoolYear = SafeDb.SafeString(Row["idSchoolYear"]);
+            Class.UriWebApp = SafeDb.SafeString(Row["uriWebApp"]);
+            Class.Description = SafeDb.SafeString(Row["desc"]);
+        }
+
+        internal Class GetClass(string IdSchool, string IdSchoolYear, string ClassAbbreviation)
+        {
+            Class c = new Class();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd = conn.CreateCommand();
+                string query = "SELECT Classes.*" +
+                   " FROM Classes" +
+                   " WHERE" +
+                   " Classes.idSchoolYear = '" + SqlVal.SqlString(IdSchoolYear) + "'" +
+                   " AND Classes.abbreviation = '" + SqlVal.SqlString(ClassAbbreviation) + "'";
+                if (IdSchool != null && IdSchool != "")
+                    query += " AND Classes.idSchool = '" + SqlVal.SqlString(IdSchool) + "'";
+                query += ";";
+
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+
+                while (dRead.Read())
+                {
+                    GetClassFromRow(c, dRead);
+                    break; // just the first! 
+                }
+            }
+            return c;
+        }
+
+        internal void SaveClass(Class Class)
+        {
+            //bool leaveConnectionOpen = true;
+            //if (conn == null)
+            //{
+            //    conn = dl.Connect();
+            //    leaveConnectionOpen = false;
+            //}
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                string query = "UPDATE Classes" +
+                    " SET" +
+                    " idClass=" + Class.IdClass + "" +
+                    ",idSchoolYear='" + SqlVal.SqlString(Class.SchoolYear) + "'" +
+                    ",idSchool='" + SqlVal.SqlString(Class.IdSchool) + "'" +
+                    ",abbreviation='" + SqlVal.SqlString(Class.Abbreviation) + "'" +
+                    ",desc='" + SqlVal.SqlString(Class.Description) + "'" +
+                    ",uriWebApp='" + Class.UriWebApp + "'" +
+                    ",pathRestrictedApplication='" + SqlVal.SqlString(Class.PathRestrictedApplication) + "'" +
+                    " WHERE idClass=" + Class.IdClass +
+                    ";";
+                cmd.CommandText = query;
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+        internal List<Student> GetStudentsOfClassList(string Scuola, string Anno,
+            string SiglaClasse, bool IncludeNonActiveStudents)
+        {
+            DbDataReader dRead;
+            DbCommand cmd;
+            List<Student> ls = new List<Student>();
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT registerNumber, Classes.idSchoolYear, " +
+                               "Classes.abbreviation, Classes.idClass, Classes.idSchool, " +
+                               "Students.*" +
+                " FROM Students" +
+                " JOIN Classes_Students ON Students.idStudent=Classes_Students.idStudent" +
+                " JOIN Classes ON Classes.idClass=Classes_Students.idClass" +
+                " WHERE Classes.idSchoolYear = '" + SqlVal.SqlString(Anno) + "'" +
+                " AND Classes.abbreviation = '" + SqlVal.SqlString(SiglaClasse) + "'";
+                if (!IncludeNonActiveStudents)
+                    query += " AND (Students.disabled = 0 OR Students.disabled IS NULL)";
+                if (Scuola != null && Scuola != "")
+                    query += " AND Classes.idSchool='" + Scuola + "'";
+                query += " ORDER BY Students.LastName, Students.FirstName";
+                query += ";";
+                cmd = conn.CreateCommand();
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    Student s = GetStudentFromRow(dRead);
+                    s.Class = (string)dRead["abbreviation"];
+                    s.IdClass = (int)dRead["idClass"];
+                    ls.Add(s);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return ls;
+        }
+
+        internal Student GetStudent(int? IdStudent)
+        {
+            Student s = new Student();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT * " +
+                    "FROM Students " +
+                    "WHERE idStudent=" + IdStudent +
+                    ";";
+                dRead = cmd.ExecuteReader();
+                dRead.Read();
+                s = GetStudentFromRow(dRead);
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return s;
+        }
+
+        private Student GetStudentFromRow(DbDataReader Row)
+        {
+            Student s = new Student();
+            s.IdStudent = (int)Row["IdStudent"];
+            s.LastName = SafeDb.SafeString(Row["LastName"]);
+            s.FirstName = SafeDb.SafeString(Row["FirstName"]);
+            s.Residence = SafeDb.SafeString(Row["Residence"]);
+            s.Origin = SafeDb.SafeString(Row["Origin"]);
+            s.Email = SafeDb.SafeString(Row["Email"]);
+            if (!(Row["birthDate"] is DBNull))
+                s.BirthDate = SafeDb.SafeDateTime(Row["birthDate"]);
+            s.BirthPlace = SafeDb.SafeString(Row["birthPlace"]);
+            s.Eligible = SafeDb.SafeBool(Row["drawable"]);
+            s.Disabled = SafeDb.SafeBool(Row["disabled"]);
+            s.RevengeFactorCounter = SafeDb.SafeInt(Row["VFCounter"]);
+
+            return s;
+        }
+
+        internal Grade GetGrade(int? IdGrade)
+        {
+            Grade g = null;
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT  * FROM Grades" +
+                    " WHERE Grades.idGrade=" + IdGrade.ToString() +
+                    ";";
+                dRead = cmd.ExecuteReader();
+                dRead.Read();
+                g = GetGradeFromRow(dRead);
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return g;
+        }
+
+        private Grade GetGradeFromRow(DbDataReader Row)
+        {
+            Grade g = new Grade();
+            g.IdGrade = (int)Row["idGrade"];
+            g.IdGradeParent = SafeDb.SafeInt(Row["idGradeParent"]);
+            g.IdStudent = SafeDb.SafeInt(Row["idStudent"]);
+            g.IdGradeType = SafeDb.SafeString(Row["IdGradeType"]);
+            g.IdSchoolSubject = SafeDb.SafeString(Row["IdSchoolSubject"]);
+            //g.IdGradeTypeParent = SafeDb.SafeString(Row["idGradeTypeParent"]);
+            g.IdQuestion = SafeDb.SafeInt(Row["idQuestion"]);
+            g.Timestamp = (DateTime)Row["timestamp"];
+            g.Value = SafeDb.SafeDouble(Row["value"]);
+            g.Weight = SafeDb.SafeDouble(Row["weight"]);
+            g.CncFactor = SafeDb.SafeDouble(Row["cncFactor"]);
+            g.IdSchoolYear = SafeDb.SafeString(Row["idSchoolYear"]);
+            //g.DummyInt = (int)Row["dummyInt"]; 
+            return g;
+        }
+
+        internal void GetGradeAndStudent(Grade Grade, Student Student)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT Grades.*,Students.* FROM Grades" +
+                    " JOIN Students ON Grades.idStudent = Students.idStudent" +
+                    " WHERE Grades.idGrade=" + Grade.IdGrade.ToString() +
+                    ";";
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    Grade = GetGradeFromRow(dRead);
+                    Student = GetStudentFromRow(dRead);
+                    break; // just the first! 
+                }
+                //dRead.Dispose();
+                //cmd.Dispose();
+            }
+        }
+
+        internal void EraseGrade(int? KeyGrade)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "DELETE FROM Grades" +
+                    " WHERE idGrade=" + KeyGrade +
+                    ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+        internal List<int> GetIdStudentsNonGraded(Class Class,
+            GradeType GradeType, SchoolSubject SchoolSubject)
+        {
+            List<int> keys = new List<int>();
+
+            DbDataReader dRead;
+            DbCommand cmd;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT Classes_Students.idStudent" +
+                " FROM Classes_Students" +
+                " WHERE Classes_Students.idClass=" + Class.IdClass +
+                " AND Classes_Students.idStudent NOT IN" +
+                "(" +
+                "SELECT DISTINCT Classes_Students.idStudent" +
+                " FROM Classes_Students" +
+                " LEFT JOIN Grades ON Classes_Students.idStudent = Grades.idStudent" +
+                " WHERE Classes_Students.idClass=" + Class.IdClass +
+                " AND Grades.idSchoolSubject='" + SchoolSubject.IdSchoolSubject + "'" +
+                " AND Grades.idGradeType='" + GradeType.IdGradeType + "'" +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                ")" +
+                ";";
+                cmd = conn.CreateCommand();
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    keys.Add((int)SafeDb.SafeInt(dRead["idStudent"]));
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return keys;
+        }
+
+        internal Grade LastGradeOfStudent(Student Student, string IdSchoolYear,
+            SchoolSubject SchoolSubject, string IdGradeType)
+        {
+            DbDataReader dRead;
+            DbCommand cmd;
+            Grade g = new Grade();
+            using (DbConnection conn = dl.Connect())
+            {
+                cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT Grades.* FROM Grades" +
+                    " WHERE Grades.idStudent=" + Student.IdStudent.ToString() +
+                    " AND Grades.idSchoolSubject='" + SchoolSubject.IdSchoolSubject + "'" +
+                    " AND Grades.idGradeType='" + IdGradeType + "'" +
+                    " AND Grades.idSchoolYear='" + IdSchoolYear + "'" +
+                    " ORDER BY Grades.timestamp DESC;";
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    g = GetGradeFromRow(dRead);
+                    break; // just the first! 
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return g;
+        }
+
+        /// <summary>
+        /// Gets the number of microquestions that haven't yet a global grade
+        /// </summary>
+        /// <param Name="id"></param>
+        /// <returns></returns>
+        internal List<Grade> CountNonClosedMicroGrades(Class Class, GradeType GradeType)
+        {
+            DbDataReader dRead;
+            DbCommand cmd;
+            List<Grade> ls = new List<Grade>();
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT Grades.idStudent, Count(*) as nGrades FROM Grades," +
+                    "Grades AS Parents,Classes_Students" +
+                    " WHERE Classes_Students.idStudent = Grades.idStudent" +
+                    " AND Classes_Students.idClass =" + Class.IdClass.ToString() +
+                    " AND Grades.idGradeType = '" + GradeType.IdGradeType + "'" +
+                    " AND Parents.idGradeType = '" + GradeType.IdGradeTypeParent + "'" +
+                    " AND Grades.idGradeParent = Parents.idGrade" +
+                    " AND Parents.Value is null or Parents.Value = 0" +
+                    " GROUP BY Grades.idStudent;";
+                cmd = new SQLiteCommand(query);
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    Grade g = new Grade();
+                    g.IdStudent = (int)dRead["idStudent"];
+                    g.DummyInt = (int)dRead["nGrades"];
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return ls;
+        }
+
+
+        internal DataTable GetGradesOfClass(Class Class,
+             string IdGradeType, string IdSchoolSubject,
+             DateTime DateFrom, DateTime DateTo)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT Grades.idGrade,datetime(Grades.timeStamp),Students.idStudent," +
+                "lastName,firstName," +
+                "Grades.value AS 'grade',Grades.weight," +
+                "Grades.idGradeParent" +
+                " FROM Grades" +
+                " JOIN Students" +
+                " ON Students.idStudent=Grades.idStudent" +
+                " JOIN Classes_Students" +
+                " ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass =" + Class.IdClass +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                " AND Grades.idGradeType = '" + IdGradeType + "'" +
+                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
+                " AND Grades.Value > 0" +
+                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
+                " ORDER BY lastName, firstName, Students.idStudent, Grades.timestamp Desc;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        internal List<Couple> GetGradesOldestInClass(Class Class,
+            GradeType GradeType, SchoolSubject SchoolSubject)
+        {
+            List<Couple> couples = new List<Couple>();
+            using (DbConnection conn = dl.Connect())
+            {
+                DbDataReader dRead;
+                DbCommand cmd = conn.CreateCommand();
+                string query;
+                query = "SELECT Classes_Students.idStudent" +
+                ",MAX(timestamp) AS InstantLastQuestion" +
+                " FROM Classes_Students LEFT JOIN Grades" +
+                " ON Classes_Students.idStudent = Grades.idStudent" +
+                " WHERE Classes_Students.idClass =" + Class.IdClass +
+                " AND Grades.idGradeType = '" + GradeType.IdGradeType + "'" +
+                " AND Grades.idSchoolSubject='" + SchoolSubject.IdSchoolSubject + "'" +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                " OR Grades.idGrade IS NULL" + // takes those that haven't any grades
+                " GROUP BY Classes_Students.idStudent" +
+                " ORDER BY InstantLastQuestion DESC;"; // ???? DESC o no 
+                cmd.CommandText = query;
+                dRead = cmd.ExecuteReader();
+                while (dRead.Read())
+                {
+                    Couple c = new Couple();
+                    // we give back also the nulls, as Nows
+                    DateTime now = System.DateTime.Now;
+                    c.Key = (int)dRead["IdStudent"];
+                    if (!dRead.IsDBNull(1))
+                        c.Value = SafeDb.SafeDateTime(dRead["InstantLastQuestion"]);
+                    else
+                        c.Value = now;
+                    couples.Add(c);
+                }
+                dRead.Dispose();
+                cmd.Dispose();
+            }
+            return couples;
+        }
+
+        internal object GetWeightedAveragesOfStudent(Student currentStudent, string stringKey1, string stringKey2, DateTime value1, DateTime value2)
+        {
+            //throw new NotImplementedException();
+            return null;
+        }
+
+        internal object GetGradesWeightedAveragesOfStudent(Student currentStudent, string stringKey1, string stringKey2, DateTime value1, DateTime value2)
+        {
+            throw new NotImplementedException();
+        }
+
+        internal DataTable GetWeightedAveragesOfClass(Class Class,
+            string IdGradeType, string IdSchoolSubject, DateTime DateFrom, DateTime DateTo)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT Grades.idGrade,Students.idStudent,lastName,firstName" +
+                ",SUM(weight)/100 AS 'GradesFraction', 1 - SUM(weight)/100 AS LeftToCloseAssesments" +
+                ",COUNT() AS 'GradesCount'" +
+                " FROM Classes_Students" +
+                " LEFT JOIN Grades ON Students.idStudent=Grades.idStudent" +
+                " JOIN Students ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass =" + Class.IdClass +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                " AND (Grades.idGradeType='" + IdGradeType + "'" +
+                " OR Grades.idGradeType IS NULL)" +
+                " AND Grades.idSchoolSubject='" + IdSchoolSubject + "'" +
+                " AND Grades.value IS NOT NULL AND Grades.value <> 0" +
+                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
+                " GROUP BY Students.idStudent" +
+                " ORDER BY GradesFraction ASC, lastName, firstName, Students.idStudent;";
+                // !!!! TODO change the query to include at first rows also those students that have no grades !!!! 
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        internal DataTable GetGradesWeightedAveragesOfClass(Class Class, string IdGradeType,
+            string IdSchoolSubject, DateTime DateFrom, DateTime DateTo)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT Grades.idGrade, Students.idStudent,lastName,firstName," +
+                " SUM(Grades.value * Grades.weight)/SUM(Grades.weight) AS 'Weighted average'" +
+                // weighted RMS (Root Mean Square) as defined here: 
+                // https://stackoverflow.com/questions/10947180/weighted-standard-deviation-in-sql-server-without-aggregation-error
+                ",SQRT(SUM(Grades.weight * SQUARE(Grades.value)) / SUM(Grades.weight) - SQUARE(SUM(Grades.weight  * Grades.value) / SUM(Grades.weight))) AS 'Weighted RMS'" +
+                ",COUNT() 'Grades Count'" +
+                " FROM Grades" +
+                " JOIN Students" +
+                " ON Students.idStudent=Grades.idStudent" +
+                " JOIN Classes_Students" +
+                " ON Classes_Students.idStudent=Students.idStudent" +
+                " WHERE Classes_Students.idClass =" + Class.IdClass +
+                " AND Grades.idSchoolYear='" + Class.SchoolYear + "'" +
+                " AND Grades.idGradeType = '" + IdGradeType + "'" +
+                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
+                " AND Grades.Value > 0" +
+                " AND Grades.Timestamp BETWEEN " + SqlVal.SqlDate(DateFrom) + " AND " + SqlVal.SqlDate(DateTo) +
+                " GROUP BY Students.idStudent" +
+                " ORDER BY lastName, firstName, Students.idStudent;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        /// <summary>
+        /// Gets all the grades of a students of a specified IdGradeType that are the sons 
+        /// of another grade which has value greater than zero
+        /// </summary>
+        /// <param Name="IdStudent"></param> student's Id
+        internal DataTable GetMicroGradesOfStudentWithMacroOpen(int? IdStudent, string IdSchoolYear,
+            string IdGradeType, string IdSchoolSubject)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                // find the macro grade type of the micro grade
+                // TODO take it from a Grade passed as parameter 
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT idGradeTypeParent " +
+                    "FROM GradeTypes " +
+                    "WHERE idGradeType='" + IdGradeType + "'; ";
+                string idGradeTypeParent = (string)cmd.ExecuteScalar();
+
+                string query = "SELECT datetime(Grades.timestamp),Questions.text,Grades.value" +
+                    ",Grades.weight,Grades.IdGrade,Grades.idGradeParent,Grades.cncFactor" +
+                    ",Questions.IdQuestion,Questions.IdQuestionType,Grades.IdSchoolSubject" +
+                    ",Questions.IdTopic,Questions.Image,Questions.Duration,Questions.Difficulty" +
+                    ",Questions.*" +
+                    ",Grades.*" +
+                    " FROM Grades" +
+                    " JOIN Grades AS Parents" +
+                    " ON Grades.idGradeParent=Parents.idGrade" +
+                    " LEFT JOIN Questions" +
+                    " ON Grades.idQuestion=Questions.idQuestion" +
+                    " WHERE Grades.idStudent =" + IdStudent +
+                    " AND Grades.idSchoolYear='" + IdSchoolYear + "'" +
+                    " AND Grades.idGradeType = '" + IdGradeType + "'" +
+                    " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
+                    " AND Parents.idGradeType = '" + idGradeTypeParent + "'" +
+                    " AND Grades.idGradeParent = Parents.idGrade" +
+                    " AND (Parents.value = 0 OR Parents.value is NULL)" +
+                    " ORDER BY Grades.timestamp;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("OpenMicroGrades");
+                DAdapt.Fill(DSet);
+                DAdapt.Dispose();
+                DSet.Dispose();
+                return DSet.Tables[0];
+            }
+        }
+
+        /// <summary>
+        /// Gets all the grades of a students of a specified IdGradeType that are the sons 
+        /// of another grade which has value NOT null AND NOT equal to zero
+        /// </summary>
+        /// <param name="IdStudent"></param>
+        /// <param name="IdSchoolYear"></param>
+        /// <param name="IdGradeType"></param>
+        /// <param name="IdSchoolSubject"></param>
+        /// <returns></returns>
+        internal DataTable GetMacroGradesOfStudentClosed(int? IdStudent, string IdSchoolYear,
+            string IdGradeType, string IdSchoolSubject)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT idGrade, idStudent, value, idSchoolSubject," +
+                    "weight, cncFactor, idSchoolYear, datetime(timestamp), idGradeType, " +
+                    "idGradeParent,idQuestion" +
+                " FROM Grades" +
+                " WHERE Grades.idStudent =" + IdStudent +
+                " AND Grades.idSchoolYear='" + IdSchoolYear + "'" +
+                " AND Grades.idGradeType = '" + IdGradeType + "'" +
+                " AND Grades.idSchoolSubject = '" + IdSchoolSubject + "'" +
+                " AND Grades.Value > 0" +
+                " ORDER BY datetime(Grades.timestamp) Desc;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("ClosedMicroGrades");
+
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        internal int CreateMacroGrade(ref Grade Grade, Student Student, string IdMicroGradeType)
+        {
+            int key = NextKey("Grades", "idGrade");
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                // find the type of the macro grade of this micrograde
+                cmd.CommandText = "SELECT IdGradeTypeParent" +
+                    " FROM GradeTypes WHERE idGradetype='" + IdMicroGradeType + "';";
+                string IdMacroGrade = (string)cmd.ExecuteScalar();
+
+                // Get the Default Weight of that Grade Type
+                cmd.CommandText = "SELECT defaultWeight " +
+                    "FROM GradeTypes " +
+                    "WHERE idGradeType='" + IdMicroGradeType + "'; ";
+                double weight = (double)cmd.ExecuteScalar();
+
+                // add macrograde
+                cmd.CommandText = "INSERT INTO Grades " +
+                "(idGrade,idStudent,idGradeType,weight,cncFactor,idSchoolYear,timestamp,idSchoolSubject) " +
+                "Values (" + key + "," + Student.IdStudent +
+                ",'" + IdMacroGrade + "'" +
+                "," + SqlVal.SqlDouble(weight) + "" +
+                ",0" +
+                ",'" + Grade.IdSchoolYear + "','" +
+                System.DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss").Replace('.', ':') + "'" +
+                ",'" + Grade.IdSchoolSubject +
+                "');";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+            return key;
+        }
+
+        internal int? SaveMicroGrade(Grade Grade)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                // create a new micro assessment in grades table
+                if (Grade == null || Grade.IdGrade == null || Grade.IdGrade == 0)
+                {
+                    Grade.IdGrade = NextKey("Grades", "idGrade");
+                    cmd.CommandText = "INSERT INTO Grades " +
+                    "(idGrade, idGradeType, idGradeParent, idStudent, value, weight, " +
+                    "cncFactor,idSchoolYear, timestamp, idQuestion,idSchoolSubject) " +
+                    "Values (" + Grade.IdGrade +
+                    ",'" + SqlVal.SqlString(Grade.IdGradeType) + "'" +
+                    "," + SqlVal.SqlInt(Grade.IdGradeParent.ToString()) + "" +
+                    "," + Grade.IdStudent + "" +
+                    "," + SqlVal.SqlDouble(Grade.Value) + "" +
+                    "," + SqlVal.SqlDouble(Grade.Weight) + "" +
+                    "," + SqlVal.SqlDouble(Grade.CncFactor) + "" +
+                    ",'" + SqlVal.SqlString(Grade.IdSchoolYear) + "'" +
+                    ",'" + System.DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss").Replace('.', ':') + "'" +
+                    "," + SqlVal.SqlInt(Grade.IdQuestion.ToString()) + "" +
+                    ",'" + Grade.IdSchoolSubject + "'" +
+                    ");";
+                }
+                else
+                {
+                    cmd.CommandText = "UPDATE Grades " +
+                    "SET" +
+                    " idGrade=" + SqlVal.SqlInt(Grade.IdGrade.ToString()) + "" +
+                    ",idGradeType='" + SqlVal.SqlString(Grade.IdGradeType) + "'" +
+                    ",idGradeParent=" + SqlVal.SqlInt(Grade.IdGradeParent.ToString()) + "" +
+                    ",idStudent=" + SqlVal.SqlInt(Grade.IdStudent.ToString()) + "" +
+                    ",idSchoolYear='" + SqlVal.SqlString(Grade.IdSchoolYear) + "'" +
+                    ",timestamp='" + SqlVal.SqlString(((DateTime)Grade.Timestamp).ToString("yyyy-MM-dd HH:mm:ss")) + "'" +
+                    ",idQuestion='" + SqlVal.SqlInt(Grade.IdQuestion.ToString()) + "'" +
+                    ",idSchoolSubject='" + SqlVal.SqlString(Grade.IdSchoolSubject) + "'" +
+                    ",value=" + SqlVal.SqlDouble(Grade.Value) + "" +
+                    ",weight=" + SqlVal.SqlDouble(Grade.Weight) + "" +
+                    ",cncFactor=" + SqlVal.SqlDouble(Grade.CncFactor) + "" +
+                    " WHERE idGrade=" + SqlVal.SqlInt(Grade.IdGrade.ToString()) +
+                    ";";
+                }
+
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+            return Grade.IdGrade;
+        }
+
+        internal List<Student> GetStudentsAndSumOfWeights(Class Class,
+            GradeType GradeType, SchoolSubject SchoolSubject,
+            DateTime DateFrom, DateTime DateTo)
+        {
+            List<Student> ls = new List<Student>();
+
+            DataTable t = GetWeightedAveragesOfClass(Class, GradeType.IdGradeType,
+                SchoolSubject.IdSchoolSubject, DateFrom, DateTo);
+            foreach (DataRow row in t.Rows)
+            {
+                Student s = GetStudent((int)row["idStudent"]);
+                s.Sum = SafeDb.SafeDouble(row["GradesFraction"]);
+                s.DummyNumber = SafeDb.SafeDouble(row["GradesCount"]);
+                ls.Add(s);
+            }
+            return ls;
+        }
+
+        internal void SaveMacroGrade(int? IdStudent, int? IdParent,
+            double Grade, double Weight, string IdSchoolYear,
+            string IdSchoolSubject)
+        {
+            // !!!! TODO !!!! pass a Grade and save all the fields of a grade 
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                // creazione del macrovoto nella tabella dei voti  
+                cmd.CommandText = "UPDATE Grades " +
+                    "SET IdStudent=" + SqlVal.SqlInt(IdStudent) +
+                    ",value=" + SqlVal.SqlDouble(Grade) +
+                    ",weight=" + SqlVal.SqlDouble(Weight) +
+                    ",idSchoolYear='" + IdSchoolYear + "'" +
+                    ",idSchoolSubject='" + IdSchoolSubject +
+                    "',timestamp ='" + System.DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss").Replace('.', ':') + "' " +
+                    "WHERE idGrade = " + IdParent +
+                    ";";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+        }
+
+        internal double GetDefaultWeightOfGradeType(string IdGradeType)
+        {
+            double d;
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = "SELECT defaultWeight " +
+                    "FROM GradeTypes " +
+                    "WHERE idGradeType='" + IdGradeType + "'; ";
+                d = (double)cmd.ExecuteScalar();
+                cmd.Dispose();
+            }
+            return d;
+        }
+
+        internal DataTable GetSubGradesOfGrade(int? IdGrade)
+        {
+            DataTable t;
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT datetime(Grades.timestamp),Questions.text,Grades.value," +
+                    " Grades.weight,Grades.cncFactor,Grades.idGradeParent" +
+                    " FROM Grades" +
+                    " JOIN Questions ON Grades.idQuestion=Questions.idQuestion" +
+                    " WHERE Grades.idGradeParent =" + IdGrade +
+                    " ORDER BY Grades.timestamp;";
+
+                DataAdapter DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DataSet DSet = new DataSet("OpenMicroGrades");
+                DAdapt.Fill(DSet);
+                t = DSet.Tables[0];
+
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+            return t;
+        }
+
+        internal void CloneGrade(DataRow Riga)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                // mette peso 0 nel voto precedente  
+                cmd.CommandText = "UPDATE Grades " +
+                    " SET weight=0" +
+                    " WHERE idGrade = " + Riga["idGrade"] +
+                    ";";
+                cmd.ExecuteNonQuery();
+                // crea un nuovo voto copiato dalla riga passata
+                int codiceVoto = NextKey("Grades", "idGrade");
+
+                // aggiunge il voto copiato dalla riga passata
+                cmd.CommandText = "INSERT INTO Grades " +
+                "(idGrade,idStudent,value,weight,cncFactor,idGradeType,idGradeParent,idSchoolYear" +
+                ",timestamp,idQuestion) " +
+                "Values (" + codiceVoto + "," + Riga["idStudent"] + "," +
+                SqlVal.SqlDouble(Riga["value"]) + "," +
+                SqlVal.SqlDouble(Riga["weight"]) + ",'" +
+                SqlVal.SqlDouble(Riga["cncFactor"]) + ",'" +
+                Riga["idGradeType"] + "'," + Riga["idGradeParent"] + ",'" +
+                Riga["idSchoolYear"] + "','" +
+                System.DateTime.Now.ToString("yyyy-MM-dd HH:mm:ss").Replace('.', ':') +
+                //"'," + riga["idQuestion"].ToString() + "" +
+                "',NULL" +
+                ");";
+                cmd.ExecuteNonQuery();
+
+                // aggiusta tutte le domande figlie
+                cmd.CommandText = "UPDATE Grades " +
+                    " SET idGradeParent=" + codiceVoto +
+                    " WHERE idGradeParent = " + Riga["idGrade"] +
+                    ";";
+                cmd.ExecuteNonQuery();
+
+                cmd.Dispose();
+            }
+        }
+
+        internal void FixQuestionInGrade(int? IdGrade)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+
+                cmd.CommandText = "UPDATE Grades" +
+                           " Set" +
+                           " isFixed=TRUE" +
+                           " WHERE idGrade=" + IdGrade +
+                           ";";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+
+        }
+
+    }
+}
