+++ DataLayer/TableManage.cs
@@ -1,0 +1,122 @@
+﻿using System;
+using System.Data.Common;
+using System.Data;
+using System.Data.SQLite;
+using System.IO;
+using System.Collections.Generic;
+using gamon;
+using System.Windows.Forms;
+using SchoolGrades.DbClasses;
+
+namespace SchoolGrades.DataLayer
+{
+    //Cesare Colella, Francesco Citarella, Andrea Siboni, Riccardo Brunelli 4L
+    class TableManage
+    {
+        DataLayer dl = new DataLayer();
+
+
+        internal void SaveTableOnCvs(DataTable Table, string FileName)
+        {
+            string fileContent = "";
+            foreach (DataColumn col in Table.Columns)
+            {
+                fileContent += col.Caption + '\t';
+            }
+            fileContent += "\r\n";
+            foreach (DataRow row in Table.Rows)
+            {
+                foreach (DataColumn col in Table.Columns)
+                {
+                    fileContent += row[col].ToString() + '\t';
+                }
+                fileContent += "\r\n";
+            }
+            TextFile.StringToFile(FileName, fileContent, false);
+        }
+
+
+        public int nFieldDbDataReader(string NomeCampo, DbDataReader dr)
+        {
+            for (int i = 0; i < dr.FieldCount; i++)
+            {
+                if (dr.GetName(i) == NomeCampo)
+                {
+                    return i;
+                }
+            }
+            return -1;
+        }
+
+        internal void CompactDatabase()
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                // compact the database 
+                cmd.CommandText = "VACUUM;";
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+            //Application.Exit();
+        }
+
+        
+
+
+
+
+
+
+        internal void GetLookupTable(string Table, ref DataSet DSet, ref DataAdapter DAdapt)
+        {
+            using (DbConnection conn = dl.Connect())
+            {
+                string query = "SELECT * FROM " + Table + ";";
+                DAdapt = new SQLiteDataAdapter(query, (SQLiteConnection)conn);
+                DSet = new DataSet("OpenLookupTable");
+
+                DAdapt.Fill(DSet);
+                DAdapt.Dispose();
+                DSet.Dispose();
+            }
+        }
+
+        internal void CreateLookupTableRow(string Table, string IdTable, DataRow Row)
+        {
+            // !!!! TODO !!!! GENERALIZZARE A TABELLE CON NOMI DEI CAMPI ARBITRARI E FAR FUNZIONARE !!!!
+            string query;
+            try
+            {
+                // if key field is Integer, this works
+                int iId = (int)Row[0];
+                query = "INSERT INTO " + Table +
+                    " (" + IdTable + ", name, desc)" +
+                    " VALUES (" + iId + ",'" + Row["name"] + "','" + Row["desc"] + "'" +
+                ");";
+            }
+            catch
+            {
+                // if key field wasn't Integer, this other will work 
+                string sId = (string)Row[0];
+                query = "INSERT INTO " + Table +
+                    " (" + IdTable + ", name, desc)" +
+                    " VALUES ('" + sId + "','" + Row["name"] + "','" + Row["desc"] + "'" +
+                ");";
+            }
+            using (DbConnection conn = dl.Connect())
+            {
+                DbCommand cmd = conn.CreateCommand();
+                cmd.CommandText = query;
+
+                cmd.ExecuteNonQuery();
+                cmd.Dispose();
+            }
+        }
+
+
+
+
+
+    }
+}
